+++ BOCCHI.Common/Config/CriticalEncountersConfig.cs
@@ -1,4 +1,5 @@
 using BOCCHI.Common.Config.Fields;
+using Newtonsoft.Json;
 using Ocelot.Config;
 using Ocelot.Config.Fields;
 
@@ -9,6 +10,7 @@ namespace BOCCHI.Common.Config;
 public class CriticalEncountersConfig : IAutoConfig
 {
     [DisabledCriticalEncounterIds(Order = 0, Section = "allowlist")]
+    [JsonProperty(ObjectCreationHandling = ObjectCreationHandling.Replace)]
     public HashSet<uint> DisabledCriticalEncounterIds { get; set; } = [];
 
     public bool IsCriticalEncounterEnabled(uint criticalEncounterId) =>